Sanitizing EBS Responses This commmit introduces a major rewrite for our EBSHTTPClient service as as well as our HTTP responses. They are clean, predictable and just work
# EBSHTTPClient
I wrote as a gin middleware so that i can *lazily* respond to our client
as i parse the request. That was convenient but not good either:
`EBSHTTPClient` only responsible of sending a request to EBS services,
it is the handler that's responsible of returning the appropriate HTTP
response to the client--not EBSHTTPClient. So, it was convenient to fix
it.

Other fixes:
- Dockerfile is now working and deployed to our VPS. It uses GitHub API
to retrieve the latest commit SHA--i used this to break Docker caching
functionality