Adding Swagger 2.0 docs
This is a very major updates as we are finally adding Swagger support for noebs. It took me a lot of time to settle into API generation too, swaggo, and use it to document *all* of the APIs.

This is really huge and it will help get the project more adoption into other teams.

The other major upgrade is the separation of dashboard APIs into their package, dashboard. While this might seem quite small, it was very important. Now, the code is way more easier to read and to reason about. That will also help me to develop (and further focus) on one side of the apis without actually getting worried about the other one.