Add sentry support and increasing token timeout span
Add sentry to monitor and track errors, we have hit weird redis errors and we want to understand
what happened to trigger that.

The second commit is increasing token refresh timeout. Android minimum work manager
schedule is 15 minutes, and that is also subject to other constraints (battery optimization and alike).